+++ test/klib/test-data-buffer-print.c
@@ -50,7 +50,7 @@ rc_t run ( const char *progname )
     
     char const *const cicero = "Sed ut perspiciatis, unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam eaque ipsa, quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt, explicabo. Nemo enim ipsam voluptatem, quia voluptas sit, aspernatur aut odit aut fugit, sed quia consequuntur magni dolores eos, qui ratione voluptatem sequi nesciunt, neque porro quisquam est, qui dolorem ipsum, quia dolor sit amet consectetur adipisci[ng]velit, sed quia non-numquam [do] eius modi tempora inci[di]dunt, ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad minima veniam, quis nostrum[d] exercitationem ullam corporis suscipitlaboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum iure reprehenderit, qui inea voluptate velit esse, quam nihil molestiae consequatur, vel illum, qui dolorem eum fugiat, quo voluptas nulla pariatur?";
     size_t const orig_length = strlen(cicero);
-    int const replicount = (5000 + orig_length - 1) / orig_length; /* allocation is in 4k chunks; pick a repeat count so that the buffer has to grow at least once */
+    int const replicount = (int)( (5000 + orig_length - 1) / orig_length ); /* allocation is in 4k chunks; pick a repeat count so that the buffer has to grow at least once */
     size_t const final_length = (orig_length + 1) * replicount + 1;
     KDataBuffer buffer;
     int i;
